requirements.txt
adafruit-circuitpython-dht==3.7.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
docopt==0.6.2
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
numpy==1.23.5
pandas==1.5.2
pipreqs==0.4.11
protobuf==3.20.2
psutil==5.9.4
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
urllib3==1.26.13
Werkzeug==2.2.2
yarg==0.1.9
gunicorn==20.1.0
